dynamic postForm and deleting categories from events works

+++ logic.js
@@ -41,9 +41,9 @@ var titleInput = $("#titleInput");
 var organizerInput = $('#organizerInput');
 var statusInput = $('#statusDropdown');
 var categoryInput = $('#categoryInput');
-imageInput = $("#imageInput");
-startDate = $('#startDate');
-endDate = $('#endDate');
+var imageInput = $("#imageInput");
+var startDate = $('#startDate');
+var endDate = $('#endDate');
 $('#statusDropdown').dropdown();
 $('#categoryDropdown').dropdown();
 
@@ -176,10 +176,11 @@ function displayMonthView(month, year){
     hideDateSpans();
 
     calendarBody.empty();
-    tableContent = "";
-    daysInMonth = calculateDaysInMonth(month, year);
+    var tableContent = "";
+    var daysInMonth = calculateDaysInMonth(month, year);
     var currentDateID;
     let startDay = (new Date(year, month)).getDay();
+    var columnIDs = [];
 
     dateCounter = 1;
     for(i = 0; i < 6; i++){
@@ -200,7 +201,8 @@ function displayMonthView(month, year){
                 var columnID = calculateIDMonthView(year, month, dateCounter);
                 // console.log("columnID: " + columnID);
                 column = "<td id='" + columnID + "' class='dayCellMonthView'>";
-                column = column + "<h1>" + dateCounter + "</h1>";
+                column = column + "<h1 class='createNewEventsArea' id='createNew" + columnID + "'>" + dateCounter + "</h1>";
+                columnIDs.push(columnID);
                 dateCounter++;
             }
             else{
@@ -223,6 +225,29 @@ function displayMonthView(month, year){
     currentDateID = calculateIDMonthView(currentYear, currentMonth, currentDate);
 
     $("#" + currentDateID).addClass("currentDateStyle");
+    addClickListenersMonthView(columnIDs, year, month);
+}
+
+function addClickListenersMonthView(columnIDs, year, month){
+
+
+    columnIDs.forEach(function(columnID){
+
+        $("#createNew" + columnID).on("click", function(){
+
+            var inputString = columnID;
+            var regEx1 = /-(\d)-/;
+            var result = inputString.replace(regEx1, "-0$1-");
+            var regEx2 = /-(\d)$/;
+            result = result.replace(regEx2, "-0$1");
+
+            $('.ui.sidebar').sidebar('toggle');
+            // console.log();
+            startDate.val(result);
+            endDate.val(result);
+
+        })
+    })
 }
 
 function displayEventsMonthView(){
@@ -239,23 +264,23 @@ function displayEventsMonthView(){
     var hideEvent;
 
 
-    console.log("hideCategories: " + hideCategories)
+    // console.log("hideCategories: " + hideCategories)
 
     eventData.forEach(function(event){
 
         hideEvent = true;
         // check whether all of the events categories are in the hideCategoryArray
 
-        console.log("Event: " + event.title + " has categories: ");
+        // console.log("Event: " + event.title + " has categories: ");
 
         if(event.categories.length > 0){
             event.categories.forEach(function(category){
 
-                console.log(category.id);
-                console.log("**********************");
+                // console.log(category.id);
+                // console.log("**********************");
 
                 if(!hideCategories.includes(category.id)){
-                    console.log("hideEvent set to false");
+                    // console.log("hideEvent set to false");
                     hideEvent = false;
                 }
             })
@@ -304,10 +329,11 @@ function displayEventsMonthView(){
                 }).modal("show");
                 
             })
+            $("#editEvent" + event.id).on("click", function(){
+                alert("edit event");
+            })
 
         // }
-
-        
     })
 
 }
@@ -390,8 +416,9 @@ function displayEventsWeekView(){
                 }).modal("show");
                 
             })
-
-
+            $("#editEvent" + event.id).on("click", function(){
+                alert("edit event");
+            })
 
     })
 
@@ -519,7 +546,8 @@ function displayWeekView(minutes, hour, day, date, month, year){
     timeHeader.show();
     calendarBody.empty();
 
-    tableContent = "";
+    var tableContent = "";
+    var columnIDs = [];
 
     displayDateSpans(day, date);
 
@@ -556,9 +584,11 @@ function displayWeekView(minutes, hour, day, date, month, year){
             var columnID = setIDWeekView(j, day, date, hoursString, minutesString);
             // console.log("dateID: " + columnID);
             var column = "<td id='" + columnID + "'>";
-
-
+            column += "<h1 class='createNewEventsArea' id='createNewEventsArea" + columnID + "'></h1>"
             column = column + "</td>";
+
+            columnIDs.push(columnID);
+
             row = row + column;
 
         }
@@ -568,6 +598,41 @@ function displayWeekView(minutes, hour, day, date, month, year){
     }
 
     calendarBody.append(tableContent);
+    addClickListenersWeekView(columnIDs);
+
+}
+
+function addClickListenersWeekView(columnIDs){
+
+    var date;
+    var time;
+    var splitIndex;
+    columnIDs.forEach(function(columnID){
+
+        
+
+        $("#createNewEventsArea" + columnID).on("click", function(){
+
+            splitIndex = columnID.indexOf("T");
+            date = columnID.substring(0, splitIndex);
+            time = columnID.substring(splitIndex+1);
+            time = time.substring(0, 2) + ":" + time.substring(2, 4);
+            console.log("date: " + date);
+            console.log("time: " + time);
+
+            var inputString = date;
+            var regEx1 = /-(\d)-/;
+            var result = inputString.replace(regEx1, "-0$1-");
+            var regEx2 = /-(\d)$/;
+            result = result.replace(regEx2, "-0$1");
+
+            $('.ui.sidebar').sidebar('toggle');
+            console.log(columnID);
+            startDate.val(result);
+            endDate.val(result);
+            startTimeInput.val(time);
+        })
+    })
 
 }
 
@@ -737,7 +802,7 @@ function generateEventDIV(event, hideEvent){
         htmlString = '<div id="' + event.id + '">';
         // htmlString += '<h3>' + eventDate + '</h3>';
         htmlString += deleteSpan;
-        htmlString += '<p>' + event.title + '</p>';
+        htmlString += '<p id="editEvent' + event.id + '">' + event.title + '</p>';
 
         htmlString += '</div>';
     }
@@ -833,8 +898,6 @@ function displayCategories(){
 
 }
 
-
-
 function postCategory(){
 
     categoryPostButton.on("click", function(){
@@ -863,22 +926,89 @@ function postCategory(){
 
 function deleteCategory(categoryID){
 
-    $.ajax({
-        type: "DELETE",
-        url: "https://dhbw.cheekbyte.de/calendar/500/categories/" + categoryID,
-        success: function(){
-            console.log("Successfully deleted category");
-        },
-      }).done(function(response){
-          console.log(response);
-          deleteCategoryFromEvents();
-          loadData();
-      })
+    Promise.all(deleteCategoryFromEvents(categoryID)).then(function(){
+        console.log("promises resolved");
+        $.ajax({
+            type: "DELETE",
+            url: "https://dhbw.cheekbyte.de/calendar/500/categories/" + categoryID,
+            success: function(){
+                console.log("Successfully deleted category");
+            },
+          }).done(function(response){
+              console.log(response);
+              loadData();
+          })
+      });
+
+    
 
 }
 
-function deleteCategoryFromEvents(){
+function deleteCategoryFromEvents(categoryID){
+
+//     var promises = [];
+
+//     array.forEach(function(element) {
+//         promises.push(
+//             developer.getResources(element)
+//                 .then((data) = > {
+//                     name = data.items[0];
+//                     return developer.getResourceContent(element, file);
+//                 })
+//                 .then((response) = > {
+//                     fileContent = atob(response.content);
+//                     self.files.push({
+//                         fileName: fileName,
+//                         fileType: fileType,
+//                         content: fileContent
+//                     });
+//                 }).catch ((error) = > {
+//                     console.log('Error: ', error);
+//                 })
+//         );
+//     });
+
+// Promise.all(promises).then(() => 
+//     self.resultingFunction(self.files)
+// );
+
+
+var promises = [];
+
+    eventData.forEach(function(event){
+        if(eventContainsCategory(event, categoryID)){
+            promises.push(
+                $.ajax({
+                    type: "DELETE",
+                    url: "https://dhbw.cheekbyte.de/calendar/500/categories/" + categoryID + "/" + event.id,
+                    success: function(){
+                        console.log("Successfully deleted category from event: " + event.title);
+                    },
+                })
+            );
+        }
+        else{
+            console.log("Category not contained in event: " + event.title);
+        }
+        
+    })
+
+    return promises
     // to be implemented
+}
+
+function eventContainsCategory(event, categoryID){
+
+    var containsCategory = false;
+
+    event.categories.forEach(function(category){
+        if(category.id === categoryID){
+            containsCategory = true;
+            console.log(event.title + " contains category with id: " + categoryID);
+        }
+    })
+
+    return containsCategory;
 }
 
 // HTML creators
@@ -946,111 +1076,109 @@ function postFormListeners(){
     titleInput.change(function() {
     getAndCheckTitleInput();
 
-  })
+    })
   
     organizerInput.change(function() {
-  checkAndGetOrganizerInput();
-})
+    checkAndGetOrganizerInput();
+    })
   
-   startTimeInput.change(function() {
-   getAndCheckStartInput();
- })
+    startTimeInput.change(function() {
+    getAndCheckStartInput();
+    })
   
-  endTimeInput.change(function() {
-  getAndCheckEndInput();
-})
+    endTimeInput.change(function() {
+    getAndCheckEndInput();
+    })
   
       statusInput.change(function() {
     getStatusInput();
-  })
+    })
 
     // remove popup after mouse leaves submit button
   document.getElementById('submitButton').onmouseout = function(event) {
     submitButton.popup('destroy');
 }
   
-          submitButton.click(function() {
-
-
+submitButton.click(function() {
 
         var locationValue = document.getElementById("locationInput").value;
         var websiteValue = document.getElementById("websiteInput").value;
         var alldayValue = $('input[name=allday]').is(':checked');
 
-            // reformat time input and add it if allday is true
-            if(alldayValue) {
-              startTimeInput = "00:00";
-              endTimeInput = "23:59";
-            }
+        // reformat time input and add it if allday is true
+        if(alldayValue) {
+            startTimeInput = "00:00";
+            endTimeInput = "23:59";
+        }
 
 
-            // get and get values of all mandatory fields
-            // if not given mark them as incomplete
-            var titleValue = getAndCheckTitleInput();
-            var organizerValue = checkAndGetOrganizerInput();
-            var startTimeValue = getAndCheckStartInput();
-            var endTimeValue = getAndCheckEndInput();
-            var statusValue = getStatusInput();
+        // get and get values of all mandatory fields
+        // if not given mark them as incomplete
+        var titleValue = getAndCheckTitleInput();
+        var organizerValue = checkAndGetOrganizerInput();
+        var startTimeValue = getAndCheckStartInput();
+        var endTimeValue = getAndCheckEndInput();
+        var statusValue = getStatusInput();
 
 
-            var inputIsValid = true;
+        var inputIsValid = true;
 
-            // check if all data was entered correctly
-            if(!titleValue || !organizerValue || !startTimeValue || !endTimeValue || !statusValue) {
-              inputIsValid = false;
+        // check if all data was entered correctly
+        if(!titleValue || !organizerValue || !startTimeValue || !endTimeValue || !statusValue) {
+            inputIsValid = false;
 
-              submitButton.popup();
-              submitButton.popup('show');
-
-
-            }
-
-            // check if times are valid
-            checkTimeValidity();
+            submitButton.popup();
+            submitButton.popup('show');
 
 
-            // check if dates are valid
-            checkDateValidity();
+        }
+
+        // check if times are valid
+        checkTimeValidity();
 
 
-            // check if image is valid
-            checkImageAndGetB64();
+        // check if dates are valid
+        checkDateValidity();
 
 
-
-            // else {
-            //   console.log("Start time: " + startTimeValue);
-            //   console.log("End time: " + endTimeValue);
-            //
-            //   if(startTimeValue.subtring(0,1) <= endTimeValue.substring(3,4)) {
-            //     alert(1);
-            //   } else if(startTimeValue.subtring(0,1) <= endTimeValue.substring(3,4)) {
-            //     alert(2);
-            //   } else if(startTimeValue == endTimeValue) {
-            //     alert(3);
-            //   } else {
-            //     alert(4);
-            //     inputIsValid = false;
-            //   }
-            //   submitButton.popup();
-            //   submitButton.popup('show');
-            //
-            // }
-
-
-            // if all the necessary input is give correctly a request can be made
-            if(inputIsValid) {
+        // check if image is valid
+        checkImageAndGetB64();
 
 
 
-              // make request with data
-              var dummyRequest = '{ "title": "' + titleValue + '", "location": "' + locationValue + '", "organizer": "' + organizerValue + '", "start": "' + startTimeValue + '", "end": "' + endTimeValue + '", "status": "' + statusValue + '", "allday": ' + alldayValue + ', "webpage": "' + websiteValue + '" }';
-              console.log(dummyRequest);
-            }
+        // else {
+        //   console.log("Start time: " + startTimeValue);
+        //   console.log("End time: " + endTimeValue);
+        //
+        //   if(startTimeValue.subtring(0,1) <= endTimeValue.substring(3,4)) {
+        //     alert(1);
+        //   } else if(startTimeValue.subtring(0,1) <= endTimeValue.substring(3,4)) {
+        //     alert(2);
+        //   } else if(startTimeValue == endTimeValue) {
+        //     alert(3);
+        //   } else {
+        //     alert(4);
+        //     inputIsValid = false;
+        //   }
+        //   submitButton.popup();
+        //   submitButton.popup('show');
+        //
+        // }
 
-            // $.post("https://dhbw.cheekbyte.de/calendar/500/events", JSON.parse(requestData), function(status) {
-            //   console.log(status);
-            // });
+
+        // if all the necessary input is give correctly a request can be made
+        if(inputIsValid) {
+
+
+
+            // make request with data
+            var dummyRequest = '{ "title": "' + titleValue + '", "location": "' + locationValue + '", "organizer": "' + organizerValue + '", "start": "' + startTimeValue + '", "end": "' + endTimeValue + '", "status": "' + statusValue + '", "allday": ' + alldayValue + ', "webpage": "' + websiteValue + '" }';
+            console.log(dummyRequest);
+        }
+
+        // $.post("https://dhbw.cheekbyte.de/calendar/500/events", JSON.parse(requestData), function(status) {
+        //   console.log(status);
+        // });
 
     });
